Assignment 3/Problem_4_RAG/requirements.txt
faiss-cpu==1.13.2
huggingface_hub==1.14.0
numpy==2.4.4
pydantic==2.13.4
pydantic_core==2.46.4
pysimplegui==6.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.5.0
tokenizers==0.22.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==5.8.1
typing_extensions==4.15.0
